small transfer_page structure change
